requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
filelock==3.9.0
huggingface-hub==0.12.0
idna==3.4
numpy
packaging
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
tokenizers==0.13.2
tqdm==4.64.1
transformers
datasets
typing_extensions==4.4.0
urllib3==1.26.14
pre-commit
flake8
black
python-box
tensorflow
tensorflow-datasets
Pillow
evaluate
accelerate >= 0.12.0
python-box
farm-haystack
more-itertools
